Refactor: #15 타입스크립트 환경 설정 및 사용자 등록 페이지 타입스크립트로 변경
