+++ package.json
@@ -52,10 +52,34 @@
       "no-nested-ternary": 0,
       "no-case-declarations": 0,
       "import/prefer-default-export": 0,
-      "no-param-reassign": 0
+      "no-param-reassign": 0,
+      "no-use-before-define": "off",
+      "import/extensions": [
+        "error",
+        "ignorePackages",
+        {
+          "js": "never",
+          "mjs": "never",
+          "jsx": "never",
+          "ts": "never",
+          "tsx": "never"
+        }
+      ]
     },
     "env": {
       "browser": true
+    },
+    "settings": {
+      "import/resolver": {
+        "node": {
+          "extensions": [
+            ".js",
+            ".jsx",
+            ".ts",
+            ".tsx"
+          ] 
+        }
+      }
     }
   },
   "browserslist": {
@@ -77,5 +101,13 @@
       ".js",
       ".jsx"
     ]
+  },
+  "devDependencies": {
+    "@types/jest": "^27.5.0",
+    "@types/node": "^17.0.31",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "@types/styled-components": "^5.1.25",
+    "typescript": "^4.6.4"
   }
 }
