+++ src/app.jsx
@@ -3,7 +3,7 @@ import './app.css';
 import { Router, Switch, Route, Redirect } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
-import UserPage from './components/user';
+import UserPage from './pages/UserPage';
 import TestPage from './components/test';
 import CompletePage from './components/complete';
 import PsychologyTestResult from './components/result';
