+++ src/components/user/UserName.tsx
@@ -1,0 +1,66 @@
+import React, { useImperativeHandle, useRef } from 'react';
+import styled from 'styled-components';
+import { useDispatch, useSelector } from 'react-redux';
+import { actions, selector } from '../../store/modules';
+import { COLOR_DARKSET } from '../../variables';
+
+const StyledLabel = styled.label`
+  > input {
+    width: 12rem;
+    padding: 0.5rem 0;
+    border-style: none;
+    border-bottom: solid 2px ${COLOR_DARKSET.HIGHLIGHT_TEXT};
+    background-color: transparent;
+    color: white;
+    text-align: center;
+    font-size: 1.1rem;
+    margin-bottom: 0.3rem;
+
+    ::placeholder {
+      color: ${COLOR_DARKSET.FONT};
+      font-size: 1rem;
+      text-align: center;
+    }
+    :focus {
+      outline: none;
+      ::placeholder {
+        font-size: 0;
+      }
+    }
+  }
+`;
+
+export interface UserNameInputRef {
+  focus: () => void;
+}
+
+const UserName: React.ForwardRefRenderFunction<UserNameInputRef> = (_, ref) => {
+  const inputRef = useRef<HTMLInputElement>(null);
+  const dispatch = useDispatch();
+  const userName = useSelector(selector.getUserName);
+
+  useImperativeHandle(ref, () => ({
+    focus: () => {
+      inputRef.current?.focus();
+    },
+  }));
+
+  return (
+    <StyledLabel htmlFor="userName">
+      <input hidden type="text" />
+      <input
+        ref={inputRef}
+        id="userName"
+        defaultValue={userName}
+        placeholder="이름을 입력해주세요."
+        type="text"
+        onChange={(e) => {
+          const name = e.target.value;
+          dispatch(actions.saveName(name));
+        }}
+      />
+    </StyledLabel>
+  );
+};
+
+export default React.forwardRef<UserNameInputRef>(UserName);
